atividade crud com relações

+++ app/Http/Controllers/LivrosController.php
@@ -5,6 +5,10 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Livro;
 use App\Models\Genero;
+use App\Models\Autor;
+use App\Models\Editora;
+
+
 
 class LivrosController extends Controller
 {
@@ -26,9 +30,13 @@ class LivrosController extends Controller
         ]);
     }
     public function create(){
-               $generos = Genero::all();
+            $autores = Autor::all();
+            $generos = Genero::all();
+            $editoras = Editora::all();
         return view('livros.create',[
-            'generos'=>$generos
+            'generos'=>$generos,
+            'autores'=>$autores,
+            'editoras'=>$editoras
         ]);
     }
     
@@ -44,10 +52,13 @@ class LivrosController extends Controller
             'observacoes'=>['nullable', 'min:1', 'max:255'],
             'imagem_capa'=>['nullable', 'min:1', 'max:255'],
             'id_genero'=>['nullable', 'numeric', 'min:10'],
-            'id_autor'=>['nullable', 'numeric', 'min:10'],
             'sinopse'=>['nullable', 'min:1', 'max:255']
         ]);
-    $livro = Livro::create($novoLivro);
+        $autores = $request->id_autor;
+        $editoras=$request->id_editora;
+        $livro = Livro::create($novoLivro); 
+        $livro->autores()->attach($autores);
+        $livro->editoras()->attach($editoras);
         return redirect()->route('livros.show', [
             'id'=>$livro->id_livro
         ]);
@@ -58,10 +69,26 @@ class LivrosController extends Controller
     public function edit (Request $request){
         $id = $request->id;
         $generos = Genero::all();
+        $autores = Autor::all();
+        $editoras=Editora::all();
         $livro = livro::where('id_livro',$id)->with(['genero','autores','editoras'])->first();
+        $autoresLivros = [];
+        //obter id_autor dos autores deste livro
+        foreach($livro->autores as $autor){
+            $autoresLivros[] = $autor->id_autor;
+        }
+        $editorasLivro= [];
+        foreach($livro->editoras as $editora){
+            $editorasLivro[]=$editora->id_editora;
+        }
+
         return view('livros.edit',[
             'livro'=>$livro,
-            'generos'=>$generos
+            'generos'=>$generos,
+            'autores'=>$autores,
+            'autoresLivro'=>$autoresLivros,
+            'editoras'=>$editoras,
+            'editorasLivro'=>$editorasLivro
         ]);
         
 
@@ -70,6 +97,7 @@ class LivrosController extends Controller
     public function update (Request $request){
         $id = $request->id;
         $livro = Livro::findOrFail ($id);
+        // dd($request->all());
         $updateLivro = $request->validate([
             'titulo'=>['required', 'min:3', 'max:100'],
             'idioma'=>['required', 'min:3', 'max:10'],
@@ -78,10 +106,13 @@ class LivrosController extends Controller
             'isbn'=>['required', 'min:13', 'max:13'],
             'observacoes'=>['nullable', 'min:1', 'max:255'],
             'imagem_capa'=>['nullable', 'min:1', 'max:255'],
-            'id_genero'=>['nullable', 'numeric', 'min:10'],
+            'id_genero'=>['nullable', 'numeric'],
             'sinopse'=>['nullable', 'min:1', 'max:255']
         ]);
-        $livro->update($atualizarLivro);
+        // dd($updateLivro);
+        $autores=$request->id_autor;
+        $livro->update($updateLivro);
+        $livro->autores()->sync($autores);
         return redirect()->route('livros.show', [ 
         'id'=>$livro->id_livro
         ]);
@@ -99,9 +130,17 @@ class LivrosController extends Controller
         }
     }
     public function destroy(Request $request){
-        $idLivro = $request->id;
-        $livro = Livro::findOrFail($idLivro);
-        $livro->delete();
-        return redirect()->route('livros.index')->with('mensagem','livro eleminado!');
+        $livro= Livro::where('id_livro', $request->id)->first();
+        $autoresLivro=Livro::findOrfail($request->id)->autores;
+        $editorasLivro=Livro::findOrfail($request->id)->editoras;
+        $livro->autores()->detach($autoresLivro);
+        $livro->editoras()->detach($editorasLivro);
+        if(is_null($livro)){
+            return redirect()->route('livros.index')->with('msg','O livro não existe');
+        }
+        else{
+            $livro->delete();
+            return redirect()->route('livros.index');
+        }
     }
 }
